Update functions.image.php
This removes the forced height value from the image. With only the width set, and the max-width style set. It will force the automatic proportionally resize of the image when displayed up to the specified width. So when displayed on a small screen it will shrink the image as needed, rather than the image being partially off-screen if too large, and possibly pushing the right-hand column off the screen, and making the user have to scroll horizonally.