Update ForumHandler.php
Changes to make sure any links to post include the Topic ID. Otherwise, the notification box will not appear. It only appears when the Topic ID is part of the viewtopic.php URL. So if you are using any of the last post links, or jumping directly to a specific post, without this change, the notification box will not be present.