Update functions.image.php
Removing the height value can lead to CLS. So the better option is to add a height: auto style tag and restore the height value. This way the browser knows the size of the image ahead of time and calculates the height based on the available max-width. Thereby preventing the CLS from happening.