Update functions.image.php
Add  -auto-orient switch to image magick thumbnail command otherwise, thumbnails could get created with the incorrect orientation if the EXIF tag in the file is specifying a different one. This happens with images taken with mobile devices. The browser automatically read the EXIF tag and views the original image correctly, but without this switch, the generated thumbnails would could show up incorrectly.